Split out parsing from memory representation
