+++ libaarhusxyz/__init__.py
@@ -1,5 +1,5 @@
-from .xyz import parse
-from .xyz import dump
+from .xyzparser import parse
+from .xyzparser import dump
 from .xyz import XYZ
 
 from .sr2 import parse as parse_sr2
